# gene_bank_service/docs/swagger.yaml
definitions:
  common.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  common.SuccessResponse:
    properties:
      status:
        type: string
    type: object
  dto.FloraRequest:
    properties:
      common_name:
        description: Common name of the plant
        type: string
      description:
        description: Description of the plant
        type: string
      image_path:
        description: Image URL or file reference
        type: string
      image_url:
        description: Image URL or file reference
        type: string
      origin:
        description: Origin of the plant
        type: string
      other_details:
        additionalProperties: true
        description: Additional details about the plant
        type: object
      scientific_name:
        description: Scientific name of the plant
        type: string
      type:
        description: Type of post
        type: string
    type: object
  dto.FloraUpdateRequest:
    properties:
      common_name:
        description: Common name of the plant
        type: string
      description:
        description: Description of the plant
        type: string
      id:
        description: Unique identifier for the plant
        type: string
      image:
        description: Image data (bytes)
        items:
          type: integer
        type: array
      image_path:
        description: Image URL or file reference
        type: string
      image_url:
        description: Image URL or file reference
        type: string
      origin:
        description: Origin of the plant
        type: string
      other_details:
        additionalProperties: true
        description: Additional details about the plant
        type: object
      scientific_name:
        description: Scientific name of the plant
        type: string
      type:
        description: Type of post
        type: string
    type: object
info:
  contact:
    email: naveenrameshcud@gmail.com
    name: Naveen R
    url: https://naveen2070.github.io/portfolio
  description: This API provides endpoints for gene bank service which is a part of
    the project chimera.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Gene Bank Service API
  version: 1.0.0
paths:
  /actuator/health:
    get:
      description: Get actuator health status based on queue and RabbitMQ statuses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Get actuator health status
      tags:
      - Actuator
  /flora:
    post:
      consumes:
      - application/json
      parameters:
      - description: Flora data
        in: body
        name: flora
        required: true
        schema:
          $ref: '#/definitions/dto.FloraRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Add a flora data to the database
      tags:
      - Flora
    put:
      consumes:
      - application/json
      parameters:
      - description: Flora data
        in: body
        name: flora
        required: true
        schema:
          $ref: '#/definitions/dto.FloraUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ErrorResponse'
      summary: Update a flora data in the database
      tags:
      - Flora
swagger: "2.0"